+++ src/api/api.ts
@@ -16,6 +16,8 @@ import type {
   BackendError,
   UserData,
   QueueHealthData,
+  SparvAnalysesData,
+  SparvLanguagesData,
 } from "@/api/api.types";
 
 /** Create a `text/yaml` file object with content */
@@ -70,12 +72,28 @@ export class MinkClient {
     return response.data;
   });
 
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Documentation/operation/list-sparv-analyses */
+  sparvAnalyses = once(async () => {
+    const response = await this.axios.get<MinkResponse<SparvAnalysesData>>(
+      "corpus/sparv/list-analyses",
+    );
+    return response.data.analyses;
+  });
+
   /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Documentation/operation/list-sparv-exports */
   sparvExports = once(async () => {
     const response = await this.axios.get<MinkResponse<SparvExportsData>>(
       "corpus/sparv/list-exports",
     );
     return response.data.exports;
+  });
+
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Documentation/operation/list-sparv-languages */
+  sparvLanguages = once(async () => {
+    const response = await this.axios.get<MinkResponse<SparvLanguagesData>>(
+      "corpus/sparv/list-languages",
+    );
+    return response.data.languages;
   });
 
   /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Documentation/operation/get-sparv-schema */