+++ src/corpus/views/CorpusCreateView.vue
@@ -3,6 +3,7 @@ import { computed } from "vue";
 import { useI18n } from "vue-i18n";
 import { FormKit } from "@formkit/vue";
 import { PhLightbulbFilament } from "@phosphor-icons/vue";
+import { toLangKey, useSparv } from "../sparv.composable";
 import PageTitle from "@/components/PageTitle.vue";
 import LayoutSection from "@/components/LayoutSection.vue";
 import useSpin from "@/spin/spin.composable";
@@ -13,8 +14,11 @@ import HelpBox from "@/components/HelpBox.vue";
 import FormKitWrapper from "@/components/FormKitWrapper.vue";
 import useAlert from "@/alert/alert.composable";
 import { CORPUS_SOURCE_FORMATS } from "@/file";
+import { useAppConfig } from "@/app/useAppConfig";
 
+const { appConfig } = useAppConfig();
 const { createCorpus } = useCreateCorpus();
+const { languageOptions } = useSparv();
 const { t } = useI18n();
 const { spin } = useSpin();
 const { showAlert } = useAlert();
@@ -23,6 +27,7 @@ type Form = {
   name?: string;
   description?: string;
   format: CorpusSourceFormat;
+  language: string;
   textAnnotation?: string;
 };
 
@@ -41,6 +46,7 @@ async function submit(fields: Form) {
     fields.name?.trim() || "",
     fields.description?.trim() || "",
     fields.format,
+    fields.language,
     fields.textAnnotation,
   );
 
@@ -74,6 +80,17 @@ async function submit(fields: Form) {
             name="name"
             input-class="w-72"
             :help="$t('metadata.name.help')"
+          />
+
+          <FormKit
+            name="language"
+            :label="$t('config.language')"
+            type="select"
+            :value="toLangKey(appConfig.defaultLanguage || '')"
+            input-class="w-72"
+            :options="languageOptions"
+            validation="required"
+            :help="$t('config.language.help')"
           />
 
           <FormKit